ref : 요양보호사 근무조건 엔드포인트 통합
