+++ homecare/src/main/java/jaega/homecare/domain/caregiver/controller/CaregiverController.java
@@ -7,6 +7,9 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jaega.homecare.domain.caregiver.dto.req.CaregiverSignupRequest;
 import jaega.homecare.domain.caregiver.dto.req.ChoiceCaregiverCenterRequest;
 import jaega.homecare.domain.caregiver.dto.res.*;
+import jaega.homecare.domain.caregiverPreference.dto.req.CreateCaregiverPreferenceRequest;
+import jaega.homecare.domain.caregiverPreference.dto.req.UpdateCaregiverPreferenceRequest;
+import jaega.homecare.domain.caregiverPreference.dto.res.GetCaregiverPreferenceResponse;
 import jaega.homecare.domain.review.dto.res.CaregiverReviewItem;
 import jaega.homecare.domain.serviceMatch.dto.res.CaregiverScheduleDetailResponse;
 import jaega.homecare.domain.serviceMatch.dto.res.CaregiverScheduleResponse;
@@ -57,6 +60,29 @@ public interface CaregiverController {
     @ApiResponse(responseCode = "200", description = "마이페이지의 요양보호사 기관 소속별 정산 내역 조회 성공")
     @GetMapping("/my-page/settlement")
     ResponseEntity<List<GetCaregiverCenterSettlementResponse>> getSettlementByCaregiver(UUID caregiverId);
+
+    /**
+     *
+     * 요양보호사 근무 조건 API
+     */
+
+    @Operation(summary = "요양보호사 선호(근무) 조건 작성 API", description = "입력받은 정보로 요양보호사의 선호(근무) 조건을 생성합니다.")
+    @ApiResponse(responseCode = "204", description = "요양보호사의 선호(근무) 조건 생성 성공")
+    @PostMapping("/preference")
+    ResponseEntity<Void> createCaregiverPreference(
+            @RequestParam UUID caregiverId,
+            @RequestBody CreateCaregiverPreferenceRequest request);
+
+    @Operation(summary = "요양보호사의 선호(근무) 조건 조회 API", description = "요양보호사가 자신의 선호(근무) 조건을 조회합니다.")
+    @ApiResponse(responseCode = "200", description = "요양보호사의 선호(근무)조건 조회 성공")
+    @GetMapping("/preference")
+    ResponseEntity<GetCaregiverPreferenceResponse> getCaregiverPreferenceByCaregiver(@RequestParam UUID caregiverId);
+
+    @Operation(summary = "(마이페이지) 요양보호사 선호 조건 수정 API", description = "요양보호사가 자신의 선호(근무) 조건을 조회합니다.")
+    @ApiResponse(responseCode = "204", description = "요양보호사 선호(근무)조건 수정 성공")
+    @PutMapping("/preference")
+    ResponseEntity<Void> updateCaregiverPreference(@RequestParam UUID caregiverId,
+                                                   @RequestBody UpdateCaregiverPreferenceRequest request);
 
     /**
      *
